Value transformer test
